refactor: insert 후 id 값 반환하도록 수정

+++ src/main/java/nextstep/sessions/infrastructure/SessionRepository.java
@@ -4,7 +4,7 @@ import nextstep.sessions.domain.Session;
 
 public interface SessionRepository {
 
-    int save(Session session);
+    long save(Session session);
 
     Session findByName(String name);
